Refactor hover fill to use cssText property
Using classes to apply the 'filled square' styling
is too inconvenient if the purpose of the program is
to allow users to customize their drawings with different
features, including new colors. Because there is not a
simple way to adjust CSS class properties using JS, it is
best to use a string that holds the default styling for the
grid squares such that other styles can be appended to it
(ex. adding a new fill color) or reverted to the original
(ex. setting the property equal to the default style string).