Separate fill color assign from func. generateGrid
Function generateGrid originally held an event listener
block to assign a 'square identifier' class and fill color
to every grid square. However, in order to assign new fill colors
to the squares, the fill color functionality is better separated from
the process it takes to generate the grid itself. If a new function is
created to solely take care of changing the color of each square with
the identifier class, then implementing new features of color customization
will become easier. Thus, the assignment of the identifier class has been
taken out of the event listener block, which was removed and placed in a
new function assignFillColor that can query for all the squares with the
identifier class (works because function generateGrid runs before assignFillColor
and assigns the class) and change their hover fill color accordingly.