Add button feature to gradually darken fill
This feature was approached using event delegation.
Because each consecutively filled square needed to
be darker than the last, it meant that this feature
could not be achieved by simply grouping all the squares
together and using a loop to apply a predetermined color
fill to each square as was done with all the other features.
Having the event listener placed on the parent node rather than
the squares themselves ensures that we can access the particular
hovered square using 'e.target' via the event listener regardless
of which square it is and then apply the appropriate styling.